# docs/swagger.yaml
definitions:
  app.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/articles:
    get:
      consumes:
      - application/json
      description: 根据请求参数（如状态、标签ID）返回文章列表数据和总数
      parameters:
      - description: 文章状态
        in: query
        name: state
        type: integer
      - description: 标签ID
        in: query
        name: tag_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回文章列表和总数
          schema:
            $ref: '#/definitions/app.Response'
        "400":
          description: 参数验证失败
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/app.Response'
      summary: 获取文章列表
      tags:
      - 文章
    post:
      consumes:
      - application/json
      description: 通过传入文章的相关信息（标签ID、标题、简述、内容、创建人、状态）来新增一篇文章。
      parameters:
      - description: 标签ID
        in: query
        name: tag_id
        required: true
        type: integer
      - description: 标题
        in: query
        name: title
        required: true
        type: string
      - description: 简述
        in: query
        name: desc
        required: true
        type: string
      - description: 内容
        in: query
        name: content
        required: true
        type: string
      - description: 创建人
        in: query
        name: created_by
        required: true
        type: string
      - description: 状态
        in: query
        name: state
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功返回数据
          schema:
            $ref: '#/definitions/app.Response'
        "400":
          description: 参数验证失败
          schema:
            $ref: '#/definitions/app.Response'
        "404":
          description: 标签不存在
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/app.Response'
      summary: 新增一篇文章
      tags:
      - 文章
  /api/v1/articles/{id}:
    delete:
      consumes:
      - application/json
      description: 通过文章ID删除指定的文章
      parameters:
      - description: 文章ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回成功信息
          schema:
            $ref: '#/definitions/app.Response'
        "400":
          description: 参数验证失败
          schema:
            $ref: '#/definitions/app.Response'
        "404":
          description: 文章不存在
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/app.Response'
      summary: 删除文章
      tags:
      - 文章
    get:
      consumes:
      - application/json
      description: 根据文章ID获取文章数据
      parameters:
      - description: 文章ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回文章信息
          schema:
            $ref: '#/definitions/app.Response'
        "400":
          description: 参数验证失败
          schema:
            $ref: '#/definitions/app.Response'
        "404":
          description: 文章不存在
          schema:
            $ref: '#/definitions/app.Response'
      summary: 获取单篇文章的详细信息
      tags:
      - 文章
    put:
      consumes:
      - application/json
      description: 通过文章ID和更新的参数修改文章信息（如标签ID、标题、简述、内容、修改人、状态）
      parameters:
      - description: 文章ID
        in: path
        name: id
        required: true
        type: integer
      - description: 标签ID
        in: query
        name: tag_id
        type: integer
      - description: 标题
        in: query
        name: title
        type: string
      - description: 简述
        in: query
        name: desc
        type: string
      - description: 内容
        in: query
        name: content
        type: string
      - description: 修改人
        in: query
        name: modified_by
        required: true
        type: string
      - description: 状态
        in: query
        name: state
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回成功信息
          schema:
            $ref: '#/definitions/app.Response'
        "400":
          description: 参数验证失败
          schema:
            $ref: '#/definitions/app.Response'
        "404":
          description: 文章不存在
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/app.Response'
      summary: 修改文章
      tags:
      - 文章
  /api/v1/tags:
    get:
      consumes:
      - application/json
      description: 根据请求的参数（如标签名、状态）获取标签数据
      parameters:
      - description: 标签名称
        in: query
        name: name
        type: string
      - description: 标签状态
        in: query
        name: state
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回标签列表和总数
          schema:
            $ref: '#/definitions/app.Response'
      summary: 获取标签列表
      tags:
      - 标签
    post:
      consumes:
      - application/json
      description: 创建新的标签
      parameters:
      - description: 标签名称
        in: query
        name: name
        required: true
        type: string
      - description: 标签状态
        in: query
        name: state
        type: integer
      - description: 创建人
        in: query
        name: created_by
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回成功信息
          schema:
            $ref: '#/definitions/app.Response'
        "400":
          description: 标签不存在
          schema:
            $ref: '#/definitions/app.Response'
      summary: 新增文章标签
      tags:
      - 标签
  /api/v1/tags/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定标签
      parameters:
      - description: 标签ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回成功信息
          schema:
            $ref: '#/definitions/app.Response'
        "400":
          description: 标签不存在
          schema:
            $ref: '#/definitions/app.Response'
      summary: 删除文章标签
      tags:
      - 标签
    put:
      consumes:
      - application/json
      description: 编辑已有标签的信息
      parameters:
      - description: 标签ID
        in: path
        name: id
        required: true
        type: integer
      - description: 标签名称
        in: query
        name: name
        type: string
      - description: 标签状态
        in: query
        name: state
        type: integer
      - description: 修改人
        in: query
        name: modified_by
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回成功信息
          schema:
            $ref: '#/definitions/app.Response'
        "400":
          description: 标签不存在
          schema:
            $ref: '#/definitions/app.Response'
      summary: 修改文章标签
      tags:
      - 标签
  /api/v1/tags/import:
    post:
      parameters:
      - description: Image File
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      summary: Import Image
  /auth:
    get:
      consumes:
      - application/json
      description: 通过用户名和密码进行验证，成功后返回一个 Token，供后续请求验证使用。
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回成功信息，包含 Token
          schema:
            $ref: '#/definitions/app.Response'
        "400":
          description: 参数验证失败
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: 认证失败，用户名或密码错误
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/app.Response'
      summary: 获取授权 Token
      tags:
      - 认证
swagger: "2.0"